pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "msif-isac"
version = "0.1.0"
description = "Reproduction code for 'Does Nature Know Fourier?' (Dikmen 2026)"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Apache-2.0"}
authors = [
    {name = "İsmail Can Dikmen", email = "can.dikmen@istinye.edu.tr"}
]
keywords = [
    "neuromorphic",
    "radar",
    "ISAC",
    "spiking neural network",
    "FMCW",
    "echolocation",
    "bio-inspired"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering",
]
dependencies = [
    "numpy>=1.24",
    "scipy>=1.10",
    "matplotlib>=3.7",
    "pandas>=2.0",
    "pyyaml>=6.0",
    "tqdm>=4.65",
    "h5py>=3.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "ruff>=0.1",
]

[project.urls]
Homepage = "https://github.com/candikmen/msif-isac"
Repository = "https://github.com/candikmen/msif-isac"
Library = "https://github.com/DrCanD/dikmen-spiking-neurons"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.ruff]
line-length = 100
target-version = "py310"
